+++ packages/contracts-bedrock/contracts/test/OptimismPortal.t.sol
@@ -293,7 +293,7 @@ contract OptimismPortal_FinalizeWithdrawal_Test is Portal_Initializer {
         _outputRootProof = Types.OutputRootProof({
             version: bytes32(uint256(0)),
             stateRoot: _stateRoot,
-            withdrawerStorageRoot: _storageRoot,
+            messagePasserStorageRoot: _storageRoot,
             latestBlockhash: bytes32(uint256(0))
         });
         _proposedBlockNumber = oracle.nextBlockNumber();
@@ -432,7 +432,7 @@ contract OptimismPortal_FinalizeWithdrawal_Test is Portal_Initializer {
         Types.OutputRootProof memory outputRootProof = Types.OutputRootProof({
             version: bytes32(0),
             stateRoot: stateRoot,
-            withdrawerStorageRoot: storageRoot,
+            messagePasserStorageRoot: storageRoot,
             latestBlockhash: bytes32(0)
         });
         vm.mockCall(
@@ -494,7 +494,7 @@ contract OptimismPortal_FinalizeWithdrawal_Test is Portal_Initializer {
         Types.OutputRootProof memory outputRootProof = Types.OutputRootProof({
             version: bytes32(0),
             stateRoot: stateRoot,
-            withdrawerStorageRoot: storageRoot,
+            messagePasserStorageRoot: storageRoot,
             latestBlockhash: bytes32(0)
         });
 
@@ -555,7 +555,7 @@ contract OptimismPortal_FinalizeWithdrawal_Test is Portal_Initializer {
         Types.OutputRootProof memory proof = Types.OutputRootProof({
             version: bytes32(uint256(0)),
             stateRoot: stateRoot,
-            withdrawerStorageRoot: storageRoot,
+            messagePasserStorageRoot: storageRoot,
             latestBlockhash: bytes32(uint256(0))
         });