+++ packages/contracts-periphery/contracts/L2/L2ERC721Bridge.sol
@@ -1,26 +1,29 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.15;
 
+import { ERC721Bridge } from "../universal/op-erc721/ERC721Bridge.sol";
 import {
     CrossDomainEnabled
 } from "@eth-optimism/contracts/libraries/bridge/CrossDomainEnabled.sol";
-import {
-    OwnableUpgradeable
-} from "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import { ERC165Checker } from "@openzeppelin/contracts/utils/introspection/ERC165Checker.sol";
 import { Address } from "@openzeppelin/contracts/utils/Address.sol";
 import { L1ERC721Bridge } from "../L1/L1ERC721Bridge.sol";
 import { IOptimismMintableERC721 } from "../universal/op-erc721/IOptimismMintableERC721.sol";
 import { Semver } from "@eth-optimism/contracts-bedrock/contracts/universal/Semver.sol";
+import { Initializable } from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
 
 /**
  * @title L2ERC721Bridge
  * @notice The L2 ERC721 bridge is a contract which works together with the L1 ERC721 bridge to
- *         make it possible to transfer ERC721 tokens between Optimism and Ethereum. This contract
+ *         make it possible to transfer ERC721 tokens from Ethereum to Optimism. This contract
  *         acts as a minter for new tokens when it hears about deposits into the L1 ERC721 bridge.
  *         This contract also acts as a burner for tokens being withdrawn.
+ *         **WARNING**: Do not bridge an ERC721 that was originally deployed on Optimism. This
+ *         bridge ONLY supports ERC721s originally deployed on Ethereum. Users will need to
+ *         wait for the one-week challenge period to elapse before their Optimism-native NFT
+ *         can be refunded on L2.
  */
-contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
+contract L2ERC721Bridge is ERC721Bridge, Semver {
     /**
      * @notice Emitted when an ERC721 bridge to the other network is initiated.
      *
@@ -84,13 +87,13 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
     address public otherBridge;
 
     /**
-     * @custom:semver 0.0.1
+     * @custom:semver 1.0.0
      *
      * @param _messenger   Address of the CrossDomainMessenger on this network.
      * @param _otherBridge Address of the ERC721 bridge on the other network.
      */
     constructor(address _messenger, address _otherBridge)
-        Semver(0, 0, 1)
+        Semver(1, 0, 0)
         CrossDomainEnabled(address(0))
     {
         initialize(_messenger, _otherBridge);
@@ -101,15 +104,23 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
      * @param _otherBridge Address of the ERC721 bridge on the other network.
      */
     function initialize(address _messenger, address _otherBridge) public initializer {
+        require(_messenger != address(0), "ERC721Bridge: messenger cannot be address(0)");
+        require(_otherBridge != address(0), "ERC721Bridge: other bridge cannot be address(0)");
+
         messenger = _messenger;
         otherBridge = _otherBridge;
-
-        // Initialize upgradable OZ contracts
-        __Ownable_init();
     }
 
     /**
-     * @notice Initiates a bridge of an NFT to the caller's account on L1.
+     * @notice Initiates a bridge of an NFT to the caller's account on L1. Note that this function
+     *         can only be called by EOAs. Smart contract wallets should use the `bridgeERC721To`
+     *         function after ensuring that the recipient address on the remote chain exists. Also
+     *         note that the current owner of the token on this chain must approve this contract to
+     *         operate the NFT before it can be bridged.
+     *         **WARNING**: Do not bridge an ERC721 that was originally deployed on Optimism. This
+     *         bridge only supports ERC721s originally deployed on Ethereum. Users will need to
+     *         wait for the one-week challenge period to elapse before their Optimism-native NFT
+     *         can be refunded on L2.
      *
      * @param _localToken  Address of the ERC721 on this domain.
      * @param _remoteToken Address of the ERC721 on the remote domain.
@@ -126,8 +137,12 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
         uint32 _minGasLimit,
         bytes calldata _extraData
     ) external {
-        // Modifier requiring sender to be EOA. This check could be bypassed by a malicious
-        // contract via initcode, but it takes care of the user error we want to avoid.
+        // Modifier requiring sender to be EOA. This prevents against a user error that would occur
+        // if the sender is a smart contract wallet that has a different address on the remote chain
+        // (or doesn't have an address on the remote chain at all). The user would fail to receive
+        // the NFT if they use this function because it sends the NFT to the same address as the
+        // caller. This check could be bypassed by a malicious contract via initcode, but it takes
+        // care of the user error we want to avoid.
         require(!Address.isContract(msg.sender), "L2ERC721Bridge: account is not externally owned");
 
         _initiateBridgeERC721(
@@ -142,7 +157,13 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
     }
 
     /**
-     * @notice Initiates a bridge of an NFT to some recipient's account on L1.
+     * @notice Initiates a bridge of an NFT to some recipient's account on L1. Note that the current
+     *         owner of the token on this chain must approve this contract to operate the NFT before
+     *         it can be bridged.
+     *         **WARNING**: Do not bridge an ERC721 that was originally deployed on Optimism. This
+     *         bridge only supports ERC721s originally deployed on Ethereum. Users will need to
+     *         wait for the one-week challenge period to elapse before their Optimism-native NFT
+     *         can be refunded on L2.
      *
      * @param _localToken  Address of the ERC721 on this domain.
      * @param _remoteToken Address of the ERC721 on the remote domain.
@@ -161,6 +182,8 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
         uint32 _minGasLimit,
         bytes calldata _extraData
     ) external {
+        require(_to != address(0), "ERC721Bridge: nft recipient cannot be address(0)");
+
         _initiateBridgeERC721(
             _localToken,
             _remoteToken,
@@ -193,25 +216,27 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
         uint256 _tokenId,
         bytes calldata _extraData
     ) external onlyFromCrossDomainAccount(otherBridge) {
-        // Check the target token is compliant and verify the deposited token on L1 matches the L2
-        // deposited token representation.
-        if (
-            // slither-disable-next-line reentrancy-events
-            ERC165Checker.supportsInterface(
-                _localToken,
-                type(IOptimismMintableERC721).interfaceId
-            ) && _remoteToken == IOptimismMintableERC721(_localToken).remoteToken()
-        ) {
-            // When a deposit is finalized, we give the NFT with the same tokenId to the account
-            // on L2.
-            // slither-disable-next-line reentrancy-events
-            IOptimismMintableERC721(_localToken).mint(_to, _tokenId);
-            // slither-disable-next-line reentrancy-events
-            emit ERC721BridgeFinalized(_localToken, _remoteToken, _from, _to, _tokenId, _extraData);
-        } else {
+        try this.completeOutboundTransfer(_localToken, _remoteToken, _to, _tokenId) {
+            if (_from == otherBridge) {
+                // The _from address is the address of the remote bridge if a transfer fails to be
+                // finalized on the remote chain.
+                // slither-disable-next-line reentrancy-events
+                emit ERC721Refunded(_localToken, _remoteToken, _to, _tokenId, _extraData);
+            } else {
+                // slither-disable-next-line reentrancy-events
+                emit ERC721BridgeFinalized(
+                    _localToken,
+                    _remoteToken,
+                    _from,
+                    _to,
+                    _tokenId,
+                    _extraData
+                );
+            }
+        } catch {
             // Either the L2 token which is being deposited-into disagrees about the correct address
             // of its L1 token, or does not support the correct interface.
-            // This should only happen if there is a  malicious L2 token, or if a user somehow
+            // This should only happen if there is a malicious L2 token, or if a user somehow
             // specified the wrong L2 token address to deposit into.
             // In either case, we stop the process here and construct a withdrawal
             // message so that users can get their NFT out in some cases.
@@ -221,8 +246,9 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
                 L1ERC721Bridge.finalizeBridgeERC721.selector,
                 _remoteToken,
                 _localToken,
-                _to, // switched the _to and _from here to bounce back the deposit to the sender
-                _from,
+                address(this), // Set the new _from address to be this contract since the NFT was
+                // never transferred to the recipient on this chain.
+                _from, // Refund the NFT to the original owner on the remote chain.
                 _tokenId,
                 _extraData
             );
@@ -234,6 +260,40 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
             // slither-disable-next-line reentrancy-events
             emit ERC721BridgeFailed(_localToken, _remoteToken, _from, _to, _tokenId, _extraData);
         }
+    }
+
+    /**
+     * @notice Completes an outbound token transfer. Public function, but can only be called by
+     *         this contract. It's security critical that there be absolutely no way for anyone to
+     *         trigger this function, except by explicit trigger within this contract. Used as a
+     *         simple way to be able to try/catch any type of revert that could occur during an
+     *         ERC721 mint/transfer.
+     *
+     * @param _localToken  Address of the ERC721 on this chain.
+     * @param _remoteToken Address of the corresponding token on the remote chain.
+     * @param _to          Address of the receiver.
+     * @param _tokenId     ID of the token being deposited.
+     */
+    function completeOutboundTransfer(
+        address _localToken,
+        address _remoteToken,
+        address _to,
+        uint256 _tokenId
+    ) external onlySelf {
+        require(
+            // slither-disable-next-line reentrancy-events
+            ERC165Checker.supportsInterface(_localToken, type(IOptimismMintableERC721).interfaceId),
+            "L2ERC721Bridge: local token interface is not compliant"
+        );
+        require(
+            _remoteToken == IOptimismMintableERC721(_localToken).remoteToken(),
+            "L2ERC721Bridge: wrong remote token for Optimism Mintable ERC721 local token"
+        );
+        require(_localToken != address(this), "L2ERC721Bridge: local token cannot be self");
+
+        // When a deposit is finalized, we give the NFT with the same tokenId to the account
+        // on L2. Note that safeMint makes a callback to the _to address which is user provided.
+        IOptimismMintableERC721(_localToken).safeMint(_to, _tokenId);
     }
 
     /**
@@ -258,16 +318,13 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
         uint32 _minGasLimit,
         bytes calldata _extraData
     ) internal {
+        require(_remoteToken != address(0), "ERC721Bridge: remote token cannot be address(0)");
+
         // Check that the withdrawal is being initiated by the NFT owner
         require(
             _from == IOptimismMintableERC721(_localToken).ownerOf(_tokenId),
             "Withdrawal is not being initiated by NFT owner"
         );
-
-        // When a withdrawal is initiated, we burn the withdrawer's NFT to prevent subsequent L2
-        // usage
-        // slither-disable-next-line reentrancy-events
-        IOptimismMintableERC721(_localToken).burn(_from, _tokenId);
 
         // Construct calldata for l1ERC721Bridge.finalizeBridgeERC721(_to, _tokenId)
         // slither-disable-next-line reentrancy-events
@@ -276,6 +333,11 @@ contract L2ERC721Bridge is Semver, CrossDomainEnabled, OwnableUpgradeable {
             remoteToken == _remoteToken,
             "L2ERC721Bridge: remote token does not match given value"
         );
+
+        // When a withdrawal is initiated, we burn the withdrawer's NFT to prevent subsequent L2
+        // usage
+        // slither-disable-next-line reentrancy-events
+        IOptimismMintableERC721(_localToken).burn(_from, _tokenId);
 
         bytes memory message = abi.encodeWithSelector(
             L1ERC721Bridge.finalizeBridgeERC721.selector,