+++ packages/contracts-periphery/contracts/universal/op-erc721/IOptimismMintableERC721.sol
@@ -30,25 +30,26 @@ interface IOptimismMintableERC721 is IERC721Enumerable {
     /**
      * @notice Chain ID of the chain where the remote token is deployed.
      */
-    function remoteChainId() external returns (uint256);
+    function remoteChainId() external view returns (uint256);
 
     /**
      * @notice Address of the token on the remote domain.
      */
-    function remoteToken() external returns (address);
+    function remoteToken() external view returns (address);
 
     /**
      * @notice Address of the ERC721 bridge on this network.
      */
-    function bridge() external returns (address);
+    function bridge() external view returns (address);
 
     /**
-     * @notice Mints some token ID for a user.
+     * @notice Mints some token ID for a user, checking first that contract recipients
+     *         are aware of the ERC721 protocol to prevent tokens from being forever locked.
      *
      * @param _to      Address of the user to mint the token for.
      * @param _tokenId Token ID to mint.
      */
-    function mint(address _to, uint256 _tokenId) external;
+    function safeMint(address _to, uint256 _tokenId) external;
 
     /**
      * @notice Burns a token ID from a user.