+++ packages/replica-healthcheck/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @eth-optimism/replica-healthcheck
 
+## 1.1.9
+
+### Patch Changes
+
+- 7215f4ce: Bump ethers to 5.7.0 globally
+- Updated dependencies [7215f4ce]
+- Updated dependencies [206f6033]
+- Updated dependencies [d7679ca4]
+  - @eth-optimism/common-ts@0.6.5
+  - @eth-optimism/core-utils@0.10.1
+
 ## 1.1.8
 
 ### Patch Changes